+++ .gitattributes
@@ -1,0 +1,1 @@
+assets/dist/* binary
