+++ .gitignore
@@ -2,3 +2,6 @@
 _site
 .sass-cache
 .jekyll-metadata
+
+# node
+node_modules
