+++ _layouts/default.html
@@ -4,18 +4,16 @@
     <meta name="description" content="{{ page.description }}">
     <meta name="viewport" content="width=device-width, initial-scale=1">
 
-    <link rel="stylesheet" href="/css/main.css">
+    <link rel="stylesheet" href="/css/base.css">
+    <link rel="stylesheet" href="/assets/dist/main.css">
+
     {% for file in page.css %}
       <link rel="stylesheet" href="/css/{{ file }}.css">
     {% endfor %}
   </head>
   <body>
     <div id="app">
-      <header>
-        <header-item href="/">Home</header-item>
-        <header-item href="/about">About</header-item>
-        <header-item href="/skies-of-arcadia">Skies of Arcadia</header-item>
-      </header>
+      <the-header></the-header>
       <div id="content">
         {{ content }}
       </div>
@@ -27,10 +25,6 @@
       <script src="https://cdn.jsdelivr.net/npm/vue"></script>
     {% endif %}
 
-    {% for file in page.js %}
-      <script type="module" src="/js/{{ file }}.js"></script>
-    {% endfor %}
-
-    <script type="module" src="/js/index.js"></script>
+    <script src="/assets/dist/main.js"></script>
   </body>
 </html>
