+++ assets/src/index.js
@@ -1,0 +1,4 @@
+import TheHeader from "./components/TheHeader.vue";
+import ImgMap from "./components/ImgMap.vue";
+
+new Vue({ el: '#app', components: {TheHeader, ImgMap} });
