+++ css/skies-of-arcadia.scss
@@ -13,59 +13,3 @@ figcaption {
 
   text-align: center;
 }
-
-.img-map {
-  position: relative;
-  object-fit: contain;
-  user-select: none;
-  width: 600px;
-
-  &__map {
-    width: 100%;
-  }
-
-  canvas {
-    position: absolute;
-    left: 0;
-    top: 0;
-  }
-}
-
-.disco-box {
-  position: absolute;
-  transform: translate(-50%, -50%);
-
-  text-align: center;
-
-  cursor: pointer;
-
-  &--expanded {
-    z-index: 1;
-  }
-
-  &__point {
-    font-size: 20px;
-
-    color: white;
-  }
-
-  &__tip {
-    align-items: center;
-    display: flex;
-    flex-direction: column;
-    padding: 5px 10px;
-
-    background-color: white;
-    border: 1px solid black;
-  }
-
-  &__tip-title {
-    white-space: nowrap;
-  }
-
-  &__tip-note {
-    margin-bottom: 10px;
-
-    font-size: 15px;
-  }
-}
