--- js/header-item.js
@@ -1,11 +1,0 @@
-Vue.component('header-item', {
-  props: ['href'],
-  template: `
-    <a
-      :href="href"
-      :class="$bem()"
-    >
-      <slot></slot>
-    </a>
-  `,
-});
