--- js/skies-of-arcadia/index.js
@@ -1,1 +1,0 @@
-import './img-map.js';
