+++ skies-of-arcadia/index.html
@@ -1,7 +1,5 @@
 ---
 title: Skies of Arcadia | Dan Carpenter
-js:
-  - skies-of-arcadia/index
 css:
   - skies-of-arcadia
 ---
